base_script/subtest__stackcollapse.sh: new test
This script's output has two space-separated fields. The first is a
semicolon separated stack including the program name (from the "comm"
field) and the function names from the call stack. The second is a
count. The file is sorted according to the first field. We are testing
if record and report works and we are also checking if the number of
samples reported by the script is equal to the real number of samples.

v2 (Michael): Regexp for CNT calculation adjusted (using "\w" instead
of too generic ".").