base_script/subtest__syscall-counts-by-pid.sh: new test
This script count syscall occurences divided into groups by command and
its pid. Each command can have several syscalls and number of their
occurences printed in the table. The test tests expected regexes and if
there is a support for syscall translation also checks if some expected
syscall names are there. We also check for the equal number of syscalls
for script report and perf report.