base_script/subtest__rw-by-file.sh: new test
The script rw-by-file displays r/w activity for files read/written to
for a given program. The test checks if script record works, if script
report matches expected regexps and also, if script report prints the
right number of requested and written bytes, which seems not to work for
not only this script, but also other scripts which work with r/w. We are
also testing whether the number of samples reported by the script is the
same as the number of samples.