base_kmem: added a basic test for perf-kmem tool
The command perf-kmem traces and measures kernel memory properties. It
has two variants: 'perf kmem record <command>' to record the kmem events
of an arbitrary workload and 'perf kmem stat' to report kernel memory
statistics.

The test checks if both variants work and print appropriate results
which are checked by some regexes. The test also tests options --caller,
--alloc, --raw-ip and --page. For each of this options, there is a
simple regex check and a count check. For option --page there is also a
test which checks if the number of samples reported by the command and
the real number of samples are equal.