base_sched/test_basic.sh: new test for perf sched options
This test checks basic functionality of perf sched variants - record,
latency, map, replay and script (timehist variant will be covered in a
separate file). The test checks if perf sched commands print reasonable
messages, which are checked by regexes. In 'latency' option there are
also number checks - runtime and switches columns count should give the
number in TOTAL row and in each line average delay should be less or
equal to maximum delay. In 'replay' variant is a test for equal number
of reported nr_events in the header and in the body.