base_data: new test for perf data command
The command has currently only one option, which is 'convert'. It
converts perf data file into another format (only CTF format is
supported by now).

This test checks if convert works and if it converts the same number of
samples that was previously recorded by perf record. There is also a
check if metadata file was created. There are also tests on --all option
with the same approach.