base_kmem/test_basic.sh: use output redirection for recording
This is necessary, because otherwise, in automated runs, when stdout
is not a TTY, perf uses to put the binary data to stdout, thinking
that user might have wanted it like that. Thus, without this patch,
the test fails and perf.data is bloated in text logs...

Although perf-kmem does not offer '-o' option directly, it can be
passed to perf-record using a trick (after '--').