base_mem/test_{basic,addresses}.sh: always specify the perf.data file
As in previous commit (3e7e4944f5633a79d05f19f3043cd0d074630dcd). Also,
this test was disabled if PERFSUITE_RUN_DIR was specified (= "the user
wanted the test to produce temp files in an other directory than '.'"),
this test was skipped, since it was not able to adjust where perf.data
files should be stored in. This condition could be also removed, since
we can adjust that.

Specifying perf.data always and explicitly helps the testsuite to pass
in automated environment, where running headless with redirected stdin
and stdout could confuse perf, so it expected perf.data being piped in
via stdin...