base_sched/test_timehist.sh: always specify the perf.data file
Since in some automated environments where stdio is piped, perf by default
expects the data at stdin instead of using perf.data as it's the default
behavior when used manually from command-line.

Thus, we always need to specify the perf.data for all record/report-like
commands. Also, this is required for cases when PERFSUITE_RUN_DIR is set
in order to put the data to different directory and not spoil the current
dir. This option is not widely used by the suite developers, thus it is
quite easy to forget about that.