base_kmem/test_basic.sh: Use integer division instead of multiplication in count check
In case the result gets floored, the multiplication causes the test to fail:
206 / 3 == 68 (remainder 2)
However, 3 * 68 != 206.

Changing multiplication to integer division fixes that problem.

However, the condition needed another (), since before this commit, it was:

  if (regexp_met and equation1 or equation2)

Due to that, equation2 was evaluated even if !regexp_met, which was a hidden
problem in case of multiplication, but with division, in case the regexp didn't
match, equation2 evaluation failed on division by zero.

Changing it to:

  if (regexp_met and (equation1 or equation2))

... fixes the problem, since if !regexp_met, the equations are not evaluated.