base_config: new directory for perf config command test
With perf config command, we can manage variables in a configuration
file. This command has option --user for writing to user file or reading
it and option --system for writing to system-wide file or reading it.
There is also option --list to show current config variables, name and
value, for all sections. The tests will only use --user option, because
we do not want to change the configuration system-wide.
The test 'base_config/test_basic.sh' tests help message and --list
option.