base_probe/test_adding_kernel.sh: probed function changed
The vfs_read() is no longer reliable function for this test on some archs.
On ppc64le, it happens that it gets often expanded to multiple probe points,
which basically means that the function is sometimes called normally, but
sometimes inlined, etc. Perf searches for all the points in the code where
we can consider the function being executed. That means that after adding
probes, it looks like the following:

  # perf probe -a 'vfs_read%return $retval'
  Failed to find "vfs_read%return",
   because vfs_read is an inlined function and has no return point.
  Added new event:
    probe:vfs_read__return (on vfs_read%return with $retval)

  You can now use it in all perf tools, such as:

          perf record -e probe:vfs_read__return -aR sleep 1

  # perf probe -a 'vfs_read'
  Added new events:
    probe:vfs_read       (on vfs_read)
    probe:vfs_read_1     (on vfs_read)
    probe:vfs_read_2     (on vfs_read)
    probe:vfs_read_3     (on vfs_read)

  You can now use it in all perf tools, such as:

          perf record -e probe:vfs_read_3 -aR sleep 1

  # cat /sys/kernel/debug/tracing/kprobe_events
  r:probe/vfs_read__return _text+5370424 arg1=$retval <-- actual entry +8
  p:probe/vfs_read _text+5372812   <-- some inline point
  p:probe/vfs_read_1 _text+5361408
  p:probe/vfs_read_2 _text+5370424 <-- actual entry +8
  p:probe/vfs_read_3 _text+5372008

  # perf stat -e probe:\* -- cat /proc/cpuinfo > /dev/null

   Performance counter stats for 'cat /proc/cpuinfo':

                 0      probe:vfs_read_3
                 0      probe:vfs_read_2
                 0      probe:vfs_read_1
                 5      probe:vfs_read
                 0      probe:vfs_read__return

       0.001237610 seconds time elapsed

       0.001313000 seconds user
       0.000000000 seconds sys

The mechanism of kretprobes allow us to use kretprobe *only* in case the
function is called normally, since it is placed at the actual entry point
of the function (on ppc64le, it is offsetted by +8, does not matter here).

However, since in the testcase, our example (`cat /proc/cpuinfo`) leads
not to actual function call, but to some inlined place, kretprobe can't
be used and gives us 0 hits. Perf warns us about this risk (see the log
above after adding the kretprobe). All this is just a limitation of the
kretprobing mechanism.

Due to that, we need to pick a different function which is called, se we
can probe also its return point. Such a function is getname_flags() for
now, it seems to work correctly on all arches. However, it's quite possible
that we'll need to search for another one.