base_stat/settings.sh: Disable HW {break,watch}point test on aarch64
Although Bhupesh Sharma has posted a patchset for this to kernel ML,
it was not accepted (saying something like "the whole implementation
of perf hw breakpoints on aarch64 is not good enough and should be
dropped, and finally we don't need it, since we have kprobes...").

Dropping the testcase support too.