base_report: fixed available CPU detection again
Previous solution used to use getconf to detect the number of available
CPUs. However, it does not work correctly on PowerPC LPARs, where it
detects number of CPUs available in total on the hardware, instead of
the number of CPUs available for the given LPAR.

On a 80 CPU PowerPC box, which is divided into five LPARs per 16 CPUs,
the command shows "80" instead of correct 16.

Fixes: f5eec05873286b0c65b6c7edc8f46604f07e25ed ("base_report: fixed available CPU detection")