base_report/tui_report_exit.exp: test ways of exiting 'perf report'
The 'perf report' command can be exitted by ESC, ^C or character q. In
this test, we check if all these options work correctly. In ^C and
character q test, we just need to check whether the 'perf report'
command ended, but in ESC option there is a possibility to not exit the
'perf report' command by pressing any character apart from ENTER. So we
also try to jump out of ESC exit menu by pressing ESC or character q and
after that we try to exit 'perf report' command with ESC + ENTER.