base_script/test_header.sh: test of the perf script header
This test tests --header-only and --header option of the perf script
command. The --header option shows perf.data header with trace output
whereas the --header-only option shows only perf.data header. We test
the header against some expected regexes and in the --header test we
also check whether the output of 'perf script --header-only'
concatenated with the output of 'perf script' is equal to the output of
'perf script --header'.