base_config/test_set_report.sh: precision count fix
The previous version of this file used the `perl` command exit code to
pass the number of occurrences of certain regexp. However, when the
number was bigger than 255, the returned number was incorrect.
In this version, we first store all the occurrences of the regexp into a
file and then count the number of occurrences by the `wc` command.