base_sched/test_basic.sh: added precision count
The precision in the 'sched latency count check' test was set to a fixed
number 0.005. This caused the test to fail if there were too many tasks
in the output of the `perf sched latency` command.
We now use $PRECISION variable for comparing counted and real runtime.
The value of the variable depends on the number of tasks in the `perf
sched latency` output.