base_report/test_tui.sh: use COMMAND variable in perf-record command
The true command used in the perf-record command sometimes caused
perf-record to collect no samples at all. We should use another command
to run perf-record - for now, we picked the pwd command. It is also
easier to change the command by editing the value of the COMMAND
variable.

In some cases, the perf-record command did not collect any samples.
Therefore we use -a option to collect more samples and to collect at
least some samples during the recording.