base_probe/test_sdt.sh: Handle libpthread move to libc
Since libpthread has been merged into libc, so now it is not in
separate library. The test needs to be adjusted. The adjustment
is still not very universal, but let's hope it's sufficient for
now.

1) When caching, let's just cache libc and libpthread if exists.

2) PTHREAD_LIBRARY variable needs to be populated with the DSO where
the functions (pthread_create(), etc.) actually are. This is due to
adding the probes (putting the .so file path after the -x switch).
Let's just list the libraries in the cache and pick libpthread if
it's there (later in the alphabet) or fallback to libc.

This algorithm sucks and if it appears to be insufficient, it needs
to be replaced by something better, however, for now it might be
good-enough (TM), while not adding any further deps (e.g. a tool
for proper dumping DSOs, etc).

3) EVENTS_TO_TEST variable content is just a bit adjusted.

4) NO_OF_EVENTS_TO_TEST is derived from EVENTS_TO_TEST, instead
of running the commands twice. A cleaner solution now.