base_diff/setup.sh: record with the same frequency
In order the diff to be fair, we need to record both data sets with
the same sampling frequency. In case the sampling frequency is random,
it might get throttled by kernel in meanwhile and then the diff might
be affected.

Let's then (best-effort) try to set the max_sample_rate to 50000 and
then use 40% sample rate (20000) for both data sets.

The values seem to be quite OK on not-so-busy system and should work
fine, even if due to any reason, kernel'd tweaked the max_sample_rate
before.