base_top: creating the skeleton of perf-top command test
Creating a new directory for perf-top command. In this commit, there are
two new files added to the repository. The 'base_top/settings.sh' has a
common form with function should_support_expect_script, which detects if
the Expect command is present on the machine. This function is needed
before the testing. The next file, 'base_top/cleanup.sh', has the same
form as cleanup scripts in other 'base_*' directories.