base_top/tui_basic_exit.exp: test perf-top ability to exit
This Expect script is sourced by bash test files of base_top. This
Rxpect script tests the ability of the perf-top command to exit in
different ways. It tests the exiting with q character, exiting with ^C,
and exiting with ESC + ENTER. The expect script also tests jumping out
of the ESC window by q character and by ESC character.