base_top/tui_basic_non_existent_filter.exp: test perf-top filtering
This Expect script is sourced by bash test files of base_top. In this
Expect script, we test the perf-top filtering by symbol, which is not
present. It firstly shows the filter window by pressing the / character
and then tries to find a symbol with a non-existent name. It is expected
to not throw any error and to behave expectably.