base_stat/test_hw.sh: Enable the k+u=ku test on hybrid platforms
On hybrid platforms, basic hardware events split into multiple ones,
such as "cpu_atom/cycles/" and "cpu_core/cycles/", while on standard
platforms, "cycles" maps to a single event "cpu/cycles/".

This has to be handled in the test as well, since the above events
have to be used separately.