base_version/test_basic.sh: Extend the testcase
We wanted to see 'perf version' in the output of `perf version`, but it
is not enough.

Linux patch 81935f10e694 ("perf tools: Fix empty version number when
building outside of a git repo") fixes a bug that sometimes there is
*only* "perf version" string and nothing more.

Let's expect a number at least, so it would match the correct version
string, but fail with the trimmed one.