base_buildid/test_buildid-cache.sh: Log removed files
It happened that only 3 of expected 4 entries were successfully removed
within the removal test. We have logs of the cache state before and after
removing the entries, but we don't know which is the fourth entry that
failed to be removed. This additional log may put some light on the bug
if it happens again.