base_stat/test_intel_uncore.sh: Adjust event selection
First, `perf list uncore` should better select uncore events than just
`perf list | grep uncore`, however, we'll still keep the grep to filter
out some rubbish.

Second, the grep is adjusted to match "unc" only, since many events are
named like unc_\w+ ...

Third, the '--no-desc' switch is used to remove descriptions, but it
does not work perfectly now.

Fourth, removed manually 'uncore_frequency', since it is a metric and
needs to be used like `perf stat -M uncore_frequency` instead of with
'-e' or '--event'. This caused the test to fail on various environments.

There is '--json' switch that might be useful for us (for machine-based
parsing of events offered by `perf list`, but it seems that with the
switch, perf-list does not further accept any filtering, such as it does
like `perf list uncore`. So maybe in future...