base_sched/test_timehist.sh: Adjust regexes for tasks
The test had two flaws in regexes regarding task names. The first flaw
was that some of the regexes were not accepting symbol '.'(dot) in the
name of a task. The second flaw was caused by threads/processes with
ID -1. The test was not covering the case when thread/process had
ID consiting of [-1/positive_integer].

Properly adjust regexes to accept all possible valid tasks. Add new
regex variable to prevent any additional typos.