base_mem/test_addresses.sh: Increase the maximum for incomplete samples
The threshold for the maximum of incomplete samples was set to only 1%
of the overall number of the samples, which resulted in certain cases
that no incomplete samples could occur. After further discussion, even
though the incomplete samples should not occure regularly, the maximum
was set too low and was adjusted to 20% of the overall samples.