common/patterns.sh: Add regex for segfault
Searching for the segmentation fault line was used on multiple
occasions throughout the testsuite. It was looking for the keywords in
the segmentation fault message. An error was happening then the "core"
keyword was found in "cpu_core".

Regex was created to unify all of the searches throughout multiple files.
The bug was fixed by requiring blank charactes aroung the keyword "core".