base_record/test_basic: Fix hwcache + tracepoint fail
Test was selecting one hwcache and one tracepoint event from the list of
events. As it was set to be the first event in the list, in cases when
at least one of the events was not supported on all of the cpus, the
recording and the test failed.

Selection of the events is now set to iterate through the list of responding
type of events until supported event is found. If no such event is
found, test is skipped.