base_config/test_set_report.sh: Fix report.children variable bug
When computing report.children variable on systems with 'cpu_core' and
'cpu_atom' the event was recorded twice and the results were around twice
as high and out of limits.

The selected event was changed to softwere event to remove the problem
with different cpu types. The "cpu-clock" event was selected as the best
option as it is the most common sw event and should be available at all
systems.