base_sched/test_timehist.sh: Remove regex for terminated tasks
With patch 40826c45eb0b the notation of the dead threads was removed from
command 'perf sched' as it was done only in the best effort way.

The test still expected the terminated tasks in the summary output.
Remove the regex for terminated tasks as this feature is no longer
available with 'perf sched'.