base_sched/test_timehist.sh: Return regex for terminated tasks for rhel8
The line with terminated tasks cannot be expected for all kernels as
this feature was removed with patch 40826c45eb0b in rhel9, but it is
still expected for rhel8. Without the terminated tasks line for checking
all_lines_match the test fails for rhel9. Return the line back allow its
appearance.