common/patterns.sh: Fix regex for trace line
Multiple rexeges(RE_LINE_TRACE_FULL/ONE_PROC/CONTINUED) used for perf
trace tests contained a mistake when the 'or' operator wasn't correctly
enclosed in brackets, which resulted in that most of the times line was
matched to a hex number and expression '.*'.

Correctly fix the regexes, add cases that were missed as
RE_LINE_TRACE_UNFINISHED and fix base_trace tests that were using the
incorrect regexes.