common/patterns.sh: Add regex for returned flags
When tracing processes there is a possibility that instread of returning
a value (either a decimal or hexadecimal), the returned value is a set
of flages separated by pipe '|'. These are either referenced by their
name if it is known or as a hexadecimal value(cases that already occured).

Add a new regex RE_FLAGS for those instances and fix the trace line
regexes that were lacking and weren't able to cover these situations.