base_stat/test_pmu.sh: Fix parsing of pmu events
To correctly parse pmu events it was expected that there is a space
character between the name of the event and its description. But this is
not true if the events are so long that for the given formatting no more
spaces are added.

Fix this issue by replacing the first character of the kernel pmu event
description with space for the parsing to work correctly.