base_diff/test_bacis.sh: Fix uninitialized variable
In examples/load.c the variable 'j' was being checked sooner as it was
initialed, causing the behaviour of the test to depend on the
architecture. For most this caused that the test passed without testing
the required feature as 'j' was eaqual to zero meaning that no load was
created at all. However, the test failes even if the variable is
initialized as the testcase had wring expectations, when it required at
most 30% of load difference between 2 different recording. The load
difference is around 40-50% for all of the tested cases.

Fix the uninitialed variable and change the required difference between
2 loads to be in range 40-50%, as given the load function it would too
hard to find a small load difference.