base_kmem/test_basic.sh: Remove regex for net bytes
The net total bytes allocated report is missing if the number of bytes
freed is higher than the number of bytes allocated. However the test
expect this line for the summaryto be present, causing the test to fail.

Remove the regex for net bytes allocated as it is an optional output log.