base_config/test_set_trace.sh: Refactor used regexes
The test_set_trace contained multiple regexes that were the same or
slight alterations of the regexes already created in the
common/patterns.sh, which are used for testing the perf trace command.

Refactor the regexes used for test_set_trace, utilize the
common/patterns.sh and merge the regexes from the test case with the
generally used patterns.