base_kmem/test_basic.sh: Switch input option
Before rhel9 there were two possible syntaxes for input option for
perf kmem stat. Either with input option before the subcommand word stat
or after it. For rhel9 this ambiguity was removed leaving only the
possibility of having an input option before the subcommand. This is
also mentioned in the man page documentation:

'perf kmem [<options>] stat' to report kernel memory statistics.

Switch the input option to be before the subcommand so that it is
suitable for rhel7 to rhel9.